+++ CHANGELOG.md
@@ -7,4 +7,5 @@
 
 ### 0.7.3
 - allow ctr-c ctrl-v 
-- fixed issue with XTZ/USD price update 
+- fixed issue with XTZ/USD price update
+- fixed issue with external links